clang-tidy: fix all but one warning

+++ source/client/options_impl.cc
@@ -93,7 +93,8 @@ OptionsImpl::OptionsImpl(int argc, const char* const* argv) {
 
   // We cap on negative values. TCLAP accepts negative values which we will get here as very
   // large values. We just cap values to 2^63.
-  const uint64_t largest_acceptable_uint64_option_value = std::pow(2ull, 63ull);
+  const uint64_t largest_acceptable_uint64_option_value =
+      static_cast<uint64_t>(std::pow(2ull, 63ull));
 
   if (requests_per_second_ == 0 || requests_per_second_ > largest_acceptable_uint64_option_value) {
     throw MalformedArgvException("Invalid value for --rps");
